Updated schematic and moved files to new location
Corrected errors in the schematic which will affect the PCB. PCB not
changed yet. Updated resistor and cap values and changed amp type to
comply with the simulations. The CAD A wien-bridge PCB will not
oscillate. The relaxation oscillator will oscillate but the new values
are required. The new schematic adds new components to the circuit which
need putting into the PCB.